+++ CommonLibF4/include/RE/NetImmerse/NiTArray.h
@@ -1,6 +1,8 @@
 #pragma once
 
-#include "RE/NetImmerse/NiTCollection.h"
+#ifndef F4SE_TEST_SUITE
+	#include "RE/NetImmerse/NiTCollection.h"
+#endif
 
 namespace RE
 {
