+++ Tests/CMakeLists.txt
@@ -13,21 +13,13 @@ add_project(
 		"src/pch.h"
 )
 
-find_package(Boost MODULE REQUIRED)
 find_package(Catch2 REQUIRED CONFIG)
 
 include(Catch)
 catch_discover_tests("${PROJECT_NAME}")
 
-target_compile_definitions(
-	"${PROJECT_NAME}"
-	PRIVATE
-		BOOST_STL_INTERFACES_DISABLE_CONCEPTS
-)
-
 target_link_libraries(
 	"${PROJECT_NAME}"
 	PRIVATE
-		Boost::headers
 		Catch2::Catch2WithMain
 )
