+++ Tests/src/pch.h
@@ -93,8 +93,132 @@
 #include <vector>
 #include <version>
 
-#pragma warning(push)
-#include <boost/stl_interfaces/iterator_interface.hpp>
-#pragma warning(pop)
+#include <catch2/catch_all.hpp>
+
+namespace stl
+{
+	[[noreturn]] inline void report_and_fail(std::string_view a_msg)
+	{
+		throw std::runtime_error(std::string(a_msg));
+	}
+
+	template <class EF>                                    //
+	requires(std::invocable<std::remove_reference_t<EF>>)  //
+		class scope_exit
+	{
+	public:
+		// 1)
+		template <class Fn>
+		explicit scope_exit(Fn&& a_fn)  //
+			noexcept(std::is_nothrow_constructible_v<EF, Fn> ||
+					 std::is_nothrow_constructible_v<EF, Fn&>)  //
+			requires(!std::is_same_v<std::remove_cvref_t<Fn>, scope_exit> &&
+					 std::is_constructible_v<EF, Fn>)
+		{
+			static_assert(std::invocable<Fn>);
+
+			if constexpr (!std::is_lvalue_reference_v<Fn> &&
+						  std::is_nothrow_constructible_v<EF, Fn>) {
+				_fn.emplace(std::forward<Fn>(a_fn));
+			} else {
+				_fn.emplace(a_fn);
+			}
+		}
+
+		// 2)
+		scope_exit(scope_exit&& a_rhs)  //
+			noexcept(std::is_nothrow_move_constructible_v<EF> ||
+					 std::is_nothrow_copy_constructible_v<EF>)  //
+			requires(std::is_nothrow_move_constructible_v<EF> ||
+					 std::is_copy_constructible_v<EF>)
+		{
+			static_assert(!(std::is_nothrow_move_constructible_v<EF> && !std::is_move_constructible_v<EF>));
+			static_assert(!(!std::is_nothrow_move_constructible_v<EF> && !std::is_copy_constructible_v<EF>));
+
+			if (a_rhs.active()) {
+				if constexpr (std::is_nothrow_move_constructible_v<EF>) {
+					_fn.emplace(std::forward<EF>(*a_rhs._fn));
+				} else {
+					_fn.emplace(a_rhs._fn);
+				}
+				a_rhs.release();
+			}
+		}
+
+		// 3)
+		scope_exit(const scope_exit&) = delete;
+
+		~scope_exit() noexcept
+		{
+			if (_fn.has_value()) {
+				(*_fn)();
+			}
+		}
+
+		void release() noexcept { _fn.reset(); }
+
+	private:
+		[[nodiscard]] bool active() const noexcept { return _fn.has_value(); }
+
+		std::optional<std::remove_reference_t<EF>> _fn;
+	};
+
+	template <class EF>
+	scope_exit(EF) -> scope_exit<EF>;
+}
+
+namespace RE
+{
+	inline void* malloc(std::size_t a_bytes) { return new std::byte[a_bytes]; }
+	inline void free(void* a_ptr) { delete[] static_cast<std::byte*>(a_ptr); }
+
+	class ScrapHeap
+	{
+	public:
+		void* Allocate(std::size_t a_bytes, [[maybe_unused]] std::size_t a_alignment)
+		{
+			return malloc(a_bytes);
+		}
+
+		void Deallocate(void* a_ptr) { free(a_ptr); }
+	};
+
+	class MemoryManager
+	{
+	public:
+		[[nodiscard]] static MemoryManager& GetSingleton()
+		{
+			static MemoryManager singleton;
+			return singleton;
+		}
+
+		[[nodiscard]] ScrapHeap* GetThreadScrapHeap() { return &_scrap; }
+
+	private:
+		ScrapHeap _scrap;
+	};
+
+	template <class T>
+	class NiTMallocInterface
+	{
+	public:
+		[[nodiscard]] static T* allocate(std::size_t a_num) { return static_cast<T*>(malloc(sizeof(T) * a_num)); }
+		static void deallocate(T* a_ptr) { free(a_ptr); }
+	};
+
+	template <class T>
+	class NiTNewInterface
+	{
+	public:
+		[[nodiscard]] static T* allocate(std::size_t a_num) { return new T[a_num]; }
+		static void deallocate(T* a_ptr) { delete[] a_ptr; }
+	};
+
+	template <class T>
+	using BSCRC32 = std::hash<T>;
+
+	template <class T1, class T2>
+	using BSTTuple = std::pair<T1, T2>;
+}
 
 using namespace std::literals;
